+++ r-vm.tf
@@ -1,6 +1,6 @@
 module "bastion_vm" {
   source  = "claranet/linux-vm/azurerm"
-  version = "6.5.0"
+  version = "7.1.0"
 
   location            = var.location
   location_short      = var.location_short
@@ -61,6 +61,7 @@ module "bastion_vm" {
   os_disk_caching              = var.storage_os_disk_caching
   os_disk_custom_name          = var.storage_os_disk_custom_name
   os_disk_size_gb              = var.storage_os_disk_size_gb
+  os_disk_overwrite_tags       = var.storage_os_disk_overwrite_tags
   os_disk_storage_account_type = var.storage_os_disk_account_type
 
   # AAD SSH Login option